+++ jam/jwt/lists/__init__.py
@@ -1,7 +1,13 @@
 # -*- coding: utf-8 -*-
 
+"""
+Module for managing white and black lists.
+See: https://jam.makridenko.ru/jwt/lists/what/
+"""
+
 from .__abc_list_repo__ import ABCList
-from .list_manipulations import JSONList, RedisList
+from .json import JSONList
+from .redis import RedisList
 
 
 __all__ = ["JSONList", "RedisList", "ABCList"]
