+++ jam/jwt/lists/json.py
@@ -3,8 +3,16 @@
 import datetime
 from typing import Literal
 
-from redis import Redis
-from tinydb import Query, TinyDB
+
+try:
+    from tinydb import Query, TinyDB
+except ImportError:
+    raise ImportError(
+        """
+        No required packages found, looks like you didn't install them:
+        `pip install "jamlib[json-lists]"`
+        """
+    )
 
 from jam.jwt.lists.__abc_list_repo__ import ABCList
 
@@ -82,70 +90,3 @@ class JSONList(ABCList):
         """
         cond = Query()
         self.__list__.remove(cond.token == token)
-
-
-class RedisList(ABCList):
-    """Black/White lists in Redis, most optimal format.
-
-    Dependency required: `pip install jamlib[redis-lists]`
-
-    Attributes:
-        __list__ (Redis): Redis instance
-        exp (int | None): Token lifetime
-    """
-
-    def __init__(
-        self,
-        type: Literal["white", "black"],
-        redis_instance: Redis,
-        in_list_life_time: int | None,
-    ) -> None:
-        """Class constructor.
-
-        Args:
-            type (Literal["white", "black"]): Type og list
-            redis_instance (Redis): `redis.Redis`
-            in_list_life_time (int | None): The lifetime of a token in the list
-        """
-        super().__init__(list_type=type)
-        self.__list__ = redis_instance
-        self.exp = in_list_life_time
-
-    def add(self, token: str) -> None:
-        """Method for adding token to list.
-
-        Args:
-            token (str): Your JWT token
-
-        Returns:
-            (None)
-        """
-        self.__list__.set(name=token, value="", ex=self.exp)
-        return None
-
-    def check(self, token: str) -> bool:
-        """Method for checking if a token is present in the list.
-
-        Args:
-            token (str): Your JWT token
-
-        Returns:
-            (bool)
-        """
-        _token = self.__list__.get(name=token)
-        if not _token:
-            return False
-        else:
-            return True
-
-    def delete(self, token: str) -> None:
-        """Method for removing a token from a list.
-
-        Args:
-            token (str): Your JWT token
-
-        Returns:
-            None
-        """
-        self.__list__.delete(token)
-        return None
